+++ clap_complete/CONTRIBUTING.md
@@ -1,0 +1,11 @@
+# How to Contribute
+
+See the [clap-wide CONTRIBUTING.md](../CONTRIBUTING.md).  This will contain `clap_complete` specific notes.
+
+### Scope
+
+`clap_complete` contains the core completion generators, meaning ones
+maintained by the clap maintainers that get priority for features and fixes.
+Additional, including contributor-maintained generators can also be contributed
+to the clap repo and sit alongside `clap_complete` in a `clap_complete_<name>`
+crate.
