+++ clap_complete/src/generators/shells/bash.rs
