+++ clap_complete/src/generators/shells/fish.rs
