+++ clap_complete/src/shell.rs
