+++ clap_complete/tests/completions/bash.rs
