+++ clap_complete/tests/completions/fish.rs
