+++ clap_complete/tests/completions/powershell.rs
