+++ clap_complete/tests/completions/zsh.rs
