+++ clap_complete_fig/tests/completions/fig.rs
