+++ clap_complete_fig/tests/completions/mod.rs
@@ -1,12 +1,8 @@
 use clap::{App, AppSettings, Arg, ValueHint};
-use clap_generate::{generate, generators::*};
+use clap_complete::{generate, generators::*};
 use std::fmt;
 
-mod bash;
-mod elvish;
-mod fish;
-mod powershell;
-mod zsh;
+mod fig;
 
 #[derive(PartialEq, Eq)]
 pub struct PrettyString<'a>(pub &'a str);
