+++ clap_generate/Cargo.toml
@@ -8,7 +8,7 @@ include = [
 	"LICENSE-*",
 	"README.md"
 ]
-description = "A generator library used with clap for shell completion scripts, manpage, etc."
+description = "Renamed to clap_complete"
 repository = "https://github.com/clap-rs/clap/tree/master/clap_generate"
 documentation = "https://docs.rs/clap_generate"
 keywords = [
@@ -22,24 +22,10 @@ categories = ["command-line-interface"]
 license = "MIT OR Apache-2.0"
 readme = "README.md"
 
-[package.metadata.release]
-pre-release-replacements = [
-  {file="README.md", search="github.com/clap-rs/clap/blob/[^/]+/", replace="github.com/clap-rs/clap/blob/{{tag_name}}/", exactly=4, prerelease = true},
-]
-
-[lib]
-bench = false
-
 [dependencies]
 clap = { path = "../", version = "=3.0.0-rc.11", default-features = false, features = ["std"] }
-
-[dev-dependencies]
-pretty_assertions = "1.0"
-clap = { path = "../", version = "=3.0.0-rc.11", default-features = false, features = ["std", "derive"] }
+clap_complete = { path = "../clap_complete", version = "=3.0.0-rc.11" }
 
 [features]
 default = []
-debug = ["clap/debug"]
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
+debug = ["clap_complete/debug"]
